Inline useMouseDownOutside
