+++ src/hooks/index.ts
@@ -2,7 +2,6 @@ export * from './useCalculatedColumns';
 export * from './useFocusRef';
 export * from './useGridDimensions';
 export * from './useLatestFunc';
-export * from './useMouseDownOutside';
 export * from './useRowSelection';
 export * from './useViewportColumns';
 export * from './useViewportRows';
